+++ src/main/java/com/crypto/cryptoPortfolio/repository/ExchangeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ExchangeRepository extends JpaRepository<Exchange, Integer> {
+public interface ExchangeRepository extends JpaRepository<Exchange, Long> {
     Optional<Exchange> findByNameIgnoreCase(String name);
 }
